📝 docs(workflows): Update README and maven.yml for test & build workflow
Update the README.md to include a badge for the Maven Compile workflow,
and update the maven.yml file to change the name of the workflow to "Test & Build".
This will provide a clearer indication of the purpose of the workflow,
and make it easier for users to understand what the workflow is doing.

 benefits:
- Provides a clearer indication of the workflow's purpose
- Improves the readability of the workflow name in the GitHub Actions UI.